implementation of activities module and services

+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"clockify/activities"
 	"clockify/helpers"
 	"clockify/projects"
 	"clockify/storage"
@@ -8,6 +9,7 @@ import (
 	"clockify/users"
 	"log"
 	"os"
+	"time"
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/joho/godotenv"
@@ -38,7 +40,8 @@ func main() {
 
 	// initialization services
 	userService := users.NewUserService(db)
-	projectService := projects.NewUserService(db)
+	projectService := projects.NewProjectService(db)
+	activitiesService := activities.NewActivitiesService(db)
 
 	// register user
 	creds := types.Credentials{
@@ -102,15 +105,57 @@ func main() {
 
 	helpers.FormatMessage("Search Project Service Ended")
 
-	helpers.FormatMessage("Delete Project Service Started")
+	helpers.FormatMessage("Create Activities Service Started")
+	act := types.Activities{
+		Name:      "Second Activity",
+		StartTime: time.Now(),
+		EndTime:   time.Now().Add(2 * time.Hour),
+	}
+	act.TimeDuration = act.EndTime.Sub(act.StartTime)
+	act.UserId = user.ID
+	act.ProjectId = 1
 
-	deleteSuccessfully, err := projectService.DeleteProject(searchResult[0].Name, user.ID)
+	check, err := activitiesService.CreateActivities(act)
 	if err != nil {
-		log.Println("Error rise Delete Project ", err)
+		log.Println("Error Rise Create Activities ", err)
 	}
 
-	log.Println(deleteSuccessfully)
+	log.Println(check)
+	helpers.FormatMessage("Create Activities Service Ended")
+
+	helpers.FormatMessage("Duplicate Activities Service Started")
+
+	duplicate, err := activitiesService.DuplicateActivity(5)
+	if err != nil {
+		log.Println("Error Rise Duplicate Activities ", err)
+	}
+
+	log.Println(duplicate)
+	helpers.FormatMessage("Duplicate Activities Service Ended")
+
+	helpers.FormatMessage("Delete Project Service Started")
+
+	if len(searchResult) > 0 {
+		deleteSuccessfully, err := projectService.DeleteProject(searchResult[0].Name, user.ID)
+		if err != nil {
+			log.Println("Error rise Delete Project ", err)
+		}
+
+		log.Println(deleteSuccessfully)
+	}
+
 	helpers.FormatMessage("Delete Project Service Ended")
+
+	helpers.FormatMessage("Delete Activities Service Started")
+
+	del, err := activitiesService.DeleteActivity(2)
+	if err != nil {
+		log.Println("Error Rise Delete Activities ", err)
+	}
+
+	log.Println(del)
+
+	helpers.FormatMessage("Delete Activities Service Ended")
 
 	// migration
 	// helpers.MigrateTable(db)
